implement saving mixes
+ add: get mixes
+ handle mixes on recently played
+ move modules around to fix circular deps
